Updated -- views테스트를 하나씩 분리.
